requirements.txt
mysql-connector-python>=8.0.33
pymysql>=1.0.2
requests>=2.31.0
